fix: update UserDTO record
Updated UserDTO record to use Long instead of int on the 'id'

+++ src/main/java/es/daw2/fct_fct/dto/UserDTO.java
@@ -5,4 +5,4 @@ package es.daw2.fct_fct.dto;
  * This class is used to transfer user data between the server and client
  * without exposing valuable information.
  */
-public record UserDTO(int id, String name, String email, boolean isAdmin) {}
+public record UserDTO(Long id, String name, String email, boolean isAdmin) {}
